improve interactable path logic
